Add first time OOTB experience
localization editing, localization changing, refactor to repository pattern/interface, split article and comment database access

+++ toilet/src/java/toilet/servlet/CoronerServlet.java
@@ -5,14 +5,15 @@ import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import libOdyssey.OdysseyFilter;
+import libWebsiteTools.OdysseyFilter;
 import libWebsiteTools.imead.Local;
 
 @WebServlet(name = "CoronerServlet", description = "Error page stuff", urlPatterns = {"/coroner", "/coroner/*"})
 public class CoronerServlet extends ToiletServlet {
 
     public static final String ERROR_JSP = "/WEB-INF/Error.jsp";
-    public static final String CORONER_PREFIX = "coroner_";
+    public static final String ERROR_IFRAME_JSP = "/WEB-INF/errorIframe.jsp";
+    public static final String ERROR_PREFIX = "error_";
     private final String[] vars = {"javax.servlet.error.status_code",
         "javax.servlet.error.exception_type", "javax.servlet.error.message",
         "javax.servlet.error.exception", "javax.servlet.error.request_uri"};
@@ -24,7 +25,8 @@ public class CoronerServlet extends ToiletServlet {
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        if (null == request.getSession(false) || request.getSession().isNew()) { // trying to hack me? F U
+        if (!request.getRequestURL().toString().equals(request.getAttribute(OdysseyFilter.ORIGINAL_REQUEST_URL))
+                && (null == request.getSession(false) || request.getSession().isNew())) { // trying to hack me? F U
             OdysseyFilter.kill(request, response);
             return;
         }
@@ -44,26 +46,20 @@ public class CoronerServlet extends ToiletServlet {
         if (null == messages[0]) {
             messages[0] = "501";
         }
-
-        if (null != messages[1]) {
-            String ex = messages[1].toString();
-            if (ex.contains("DatabaseException")) {
-                messages[0] = "13";
-            } else if (ex.contains("ClassCastException")) {
-                messages[0] = "45";
-            }
-        }
         setJSPAttrs(request, response, messages[0].toString());
     }
 
     private void setJSPAttrs(HttpServletRequest request, HttpServletResponse response, String ErrCodeStr) throws ServletException, IOException {
         asyncRecentCategories(request);
-        String errorMessage = imead.getLocal(CORONER_PREFIX + ErrCodeStr, Local.resolveLocales(request));
+        String errorMessage = imead.getLocal(ERROR_PREFIX + ErrCodeStr, Local.resolveLocales(request));
         if (null == errorMessage) {
-            errorMessage = imead.getLocal(CORONER_PREFIX + "404", Local.resolveLocales(request));
+            errorMessage = imead.getLocal(ERROR_PREFIX + "404", Local.resolveLocales(request));
         }
         request.setAttribute("title", "ERROR " + ErrCodeStr);
-        request.setAttribute("mess", errorMessage);
-        getServletContext().getRequestDispatcher(ERROR_JSP).forward(request, response);
+        request.setAttribute("ERROR_MESSAGE", errorMessage);
+        try {
+            getServletContext().getRequestDispatcher(null == request.getParameter("iframe") ? ERROR_JSP : ERROR_IFRAME_JSP).forward(request, response);
+        } catch (IllegalStateException is) {
+        }
     }
 }
